BREAKING CHANGE: update responsive text widget

+++ lib/helpers/responsive_calculations.dart
@@ -69,7 +69,7 @@ class ResponsivityHelper {
   }
 
   static double responsiveFontSize(
-    int fontSize, {
+    double fontSize, {
     double? screenHeightPxReference,
     FontSizeMode? fontSizeMode,
   }) {
